move trailing restriction to hook inside BaseHeader
