+++ packages/blade/src/components/BottomSheet/utils.ts
@@ -1,6 +1,4 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
-import React from 'react';
-import { getComponentId } from '~utils';
 
 type SnapPoints = [number, number, number];
 
@@ -56,70 +54,4 @@ function computeMaxContent(
   );
 }
 
-type TrailingComponents = 'Button' | 'Badge' | 'Link' | 'Text';
-
-// prop restriction map for corresponding sub components
-const propRestrictionMap = {
-  Button: {
-    size: 'xsmall',
-    variant: 'tertiary',
-  },
-  Badge: {
-    size: 'medium',
-  },
-  Link: {
-    size: 'medium',
-  },
-  Text: {
-    size: 'medium',
-    variant: 'body',
-  },
-} as const;
-
-const useBottomSheetHeaderTrailingRestriction = (trailing: React.ReactNode) => {
-  // Could've used refs, but in react-native for some reason the validatedTrailing
-  // is changing to null on opening the bottomsheet
-  // thus saving it to state only
-  const [
-    validatedTrailingComponent,
-    setValidatedTrailingComponent,
-  ] = React.useState<React.ReactElement | null>(null);
-
-  // validate and restrict sub component props in trailing prop
-  React.useEffect(() => {
-    if (React.isValidElement(trailing)) {
-      const trailingComponentType = getComponentId(trailing) as TrailingComponents;
-      const restrictedProps = propRestrictionMap[trailingComponentType];
-      const allowedComponents = Object.keys(propRestrictionMap);
-      if (!restrictedProps) {
-        throw new Error(
-          `[Blade BottomSheetHeader]: Only one of \`${allowedComponents.join(
-            ', ',
-          )}\` component is accepted as trailing`,
-        );
-      }
-
-      const restrictedPropKeys = Object.keys(propRestrictionMap[trailingComponentType]);
-      for (const prop of restrictedPropKeys) {
-        if (trailing?.props?.hasOwnProperty(prop)) {
-          console.warn(
-            `[Blade BottomSheetHeader]: Do not pass "${prop}" to "${trailingComponentType}" while inside BottomSheetHeader trailing, because we override it.`,
-          );
-        }
-      }
-      setValidatedTrailingComponent(
-        React.cloneElement(trailing as React.ReactElement, restrictedProps),
-      );
-    }
-  }, [trailing]);
-
-  return validatedTrailingComponent;
-};
-
-export {
-  SnapPoints,
-  computeMaxContent,
-  computeMinContent,
-  computeSnapPointBounds,
-  useBottomSheetHeaderTrailingRestriction,
-};
+export { SnapPoints, computeMaxContent, computeMinContent, computeSnapPointBounds };
